Clean swagger API defs, data updates
