+++ controllers/HelpService.js
@@ -14,5 +14,6 @@ exports.helpGET = function(args, res, next) {
   else {
     res.end();
   }
-  
+
 }
+
