+++ controllers/SampleService.js
@@ -1,0 +1,68 @@
+'use strict';
+
+exports.sampledataGET = function(args, res, next) {
+  /**
+   * parameters expected in the args:
+  * growerid (String)
+  * date (String)
+  **/
+    var examples = {};
+  examples['application/json'] = {
+  "allaverage" : "",
+  "data" : [ [ {
+    "rs" : 1.3579000000000001069366817318950779736042022705078125,
+    "juicebrix" : 1.3579000000000001069366817318950779736042022705078125,
+    "sampleid" : "aeiou",
+    "moistare" : 1.3579000000000001069366817318950779736042022705078125,
+    "percentsediment" : 1.3579000000000001069366817318950779736042022705078125,
+    "maingrowerid" : "aeiou",
+    "mlmud" : 1.3579000000000001069366817318950779736042022705078125,
+    "canepol" : 1.3579000000000001069366817318950779736042022705078125,
+    "sampledate" : "2000-01-23T04:56:07.000+00:00",
+    "canebrix" : 1.3579000000000001069366817318950779736042022705078125,
+    "canefibrq" : 1.3579000000000001069366817318950779736042022705078125,
+    "labtrs" : 1.3579000000000001069366817318950779736042022705078125,
+    "id" : "",
+    "moistpercentbag" : 1.3579000000000001069366817318950779736042022705078125,
+    "seq" : "aeiou",
+    "res" : 1.3579000000000001069366817318950779736042022705078125,
+    "juicepurity" : 1.3579000000000001069366817318950779736042022705078125,
+    "polread" : 1.3579000000000001069366817318950779736042022705078125,
+    "juicepol" : 1.3579000000000001069366817318950779736042022705078125,
+    "moismd" : 1.3579000000000001069366817318950779736042022705078125,
+    "sumtotal" : 1.3579000000000001069366817318950779736042022705078125,
+    "urlpath" : "aeiou",
+    "pjbrx" : 1.3579000000000001069366817318950779736042022705078125,
+    "adjresidue" : 1.3579000000000001069366817318950779736042022705078125,
+    "adjgramsbag" : 1.3579000000000001069366817318950779736042022705078125,
+    "moismw" : 1.3579000000000001069366817318950779736042022705078125,
+    "pjpolar" : 1.3579000000000001069366817318950779736042022705078125,
+    "mud" : 1.3579000000000001069366817318950779736042022705078125,
+    "acct" : "aeiou"
+  } ] ],
+  "diff" : "",
+  "published_at" : "2000-01-23T04:56:07.000+00:00",
+  "growerid" : "aeiou",
+  "status" : "aeiou",
+  "groweraverage" : {
+    "juicepurity" : 1.3579000000000001069366817318950779736042022705078125,
+    "juicebrix" : 1.3579000000000001069366817318950779736042022705078125,
+    "growerno" : "aeiou",
+    "fibraque" : 1.3579000000000001069366817318950779736042022705078125,
+    "juicepol" : 1.3579000000000001069366817318950779736042022705078125,
+    "corelabtrs" : 1.3579000000000001069366817318950779736042022705078125,
+    "estmillcrs" : 1.3579000000000001069366817318950779736042022705078125,
+    "adjgramsbag" : 1.3579000000000001069366817318950779736042022705078125,
+    "sediment" : 1.3579000000000001069366817318950779736042022705078125
+  }
+};
+  if(Object.keys(examples).length > 0) {
+    res.setHeader('Content-Type', 'application/json');
+    res.end(JSON.stringify(examples[Object.keys(examples)[0]] || {}, null, 2));
+  }
+  else {
+    res.end();
+  }
+  
+}
+
